requirements.txt
pandas
google-generativeai
pdfplumber